Audit for error prone isinstance int/float and add lint
We recently fixed a bug on symbolic-shapes branch where
an isinstance(x, int) test failed when passed a SymIntNode.
To prevent this, I've added a lint for all the codepaths
where we may pass SymInt/SymFloat directly to reject
direct isinstance int/float tests, and instead use one of
the aliases.  The lint rule explains the options.  I then
go and fix all of them.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: c04ce735953c6a7fc9dca624fbc50ed8432927d7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87345