Simplify a few diagonal-related functions
`diag` was unnecessarily implemented as a kernel rather than as a composite
function, which made it unnecessarily difficult (explicit backward + all it entails).

We also change a few uses of `diag` on 2D tensors for `diagonal()`. The
latter returns a view rather than creating a new tensor.

We also upgrade its meta implementation to a fully-fledged
decomposition

ghstack-source-id: 7097734b5133fecdb008a4f8b7fdf2d67eda1404
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87180