[inductor] force 'fork' method for processes, cleanup
To cooperate with other multithreading methods, this
forces the process pool to use 'fork' even if others have set it
diferently. We require fork because otherwise `if __name__ == __main__`
needs to be set which we do not control as a library.

Furthermore this adds code to cleanup worker processes if
the parent exits abnormally (e.g. segfault). Previously we would leave
live but inactive workers around.

ghstack-source-id: 0b90d992f2dad524a9763b84ac481c5195635871
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87411